large-model-setups/large-model-setup-mac/requirements.txt
docx2md>=1.0.3
ipython>=9.3.0,
llama-index>=0.12.41,
qdrant-client>=1.14.2,
pymupdf4llm>=0.0.24,
llama-index-vector-stores-qdrant>=0.6.0,
llama-index-embeddings-huggingface>=0.5.4,
sentence-transformers>=4.1.0
flask
redis